v0.2.23 - Add SDCard module in WifiDAQE3AH (#59)
v0.2.23 Date: 2024/05/20, Developer: Chunglee_people
---------------------------------------------------
### Added
- New product: EthanA2
- Add SDCard, RTC_AI, encoder, counter function
- exmaple code: RTC_trigger_AI_continuous.cs, RTC_trigger_AI_N_samples,cs, RTC_trigger_AI_on_demand.cs, Encoder_read.cs, SD_read, SD_write

### Changed
- All product pinout with encoder feature
- error code table

+++ examples/Console/EthanIA/AI24Bit/AI_on_demand_once.cs
@@ -51,9 +51,7 @@ class EthanIA_AI_on_demand_once
             Console.WriteLine($"AI_open in port {port}, status: {err}");
 
             // Read data acquisition acquisition
-            List<float> sample = dev.AI_readOnDemand(port, timeout);
-
-            // Print data
+            List<double> sample = dev.AI_readOnDemand(port, timeout);
             Console.WriteLine(string.Format("[{0}]", string.Join(", ", sample)));
 
             // Close AI
